[#2131] Refactor BarChart client code
